Slightly optimize ChunkPopulator
The block removal process COULD be separated to a method that only removes blocks without any chunk data reads, by iterating through the world-space grid.
This however does not remove the need to read every block in the chunk, to determine if any special operations need to be completed for them (e.g. leaves)